tools/requirements.txt
requests
kubernetes
pycryptodome
click
